+++ package.json
@@ -10,15 +10,19 @@
     "@sb1/ffe-cards": "^12.0.4",
     "@sb1/ffe-cards-react": "^5.0.0",
     "@sb1/ffe-core": "^19.2.0",
+    "@sb1/ffe-core-react": "^7.1.4",
     "@sb1/ffe-dropdown-react": "^5.1.7",
     "@sb1/ffe-form": "^19.0.7",
     "@sb1/ffe-grid": "^12.0.0",
     "@sb1/ffe-grid-react": "^11.0.0",
     "@sb1/ffe-icons-react": "^7.2.18",
+    "@sb1/ffe-lists-react": "^7.0.5",
+    "@sb1/ffe-message-box-react": "^10.0.5",
     "@sb1/ffe-webfonts": "^3.0.0",
     "babel-jest": "^27.3.1",
     "babel-preset-gatsby": "^2.0.0",
     "classnames": "^2.3.1",
+    "clsx": "^2.0.0",
     "gatsby": "^4.15.0",
     "gatsby-image": "^3.11.0",
     "gatsby-plugin-gdpr-cookies": "^2.0.9",
