+++ src/pages/empati/components/card/text-card.module.less
@@ -2,7 +2,6 @@
 @base-color-fjell: @ffe-farge-fjell;
 @base-color-hvit: @ffe-farge-hvit;
 @base-background: @ffe-farge-fjell;
-@base-width: 950px;
 
 .card-styles(@color, @background) {
   :global(.ffe-h1),
@@ -15,7 +14,7 @@
   background: @background;
   text-align: start;
   color: @color;
-  width: @base-width;
+  border-radius: 50px;
 }
 
 :global(.ffe-text-card).colorSand {
