+++ src/pages/empati/components/card/universell-utforming.text.js
@@ -1,49 +1,49 @@
-
 const txt = {
   HEADER_STAND_1: 'Stand 1: Makuladegenerasjon (forkalkning)',
   TEXT_STAND_1:
-      'Det finnes flere typer, men aldersrelatert makuladegenerasjon er den største årsaken til varig synshemming hos personer over 50 år i Norge. \n\nMakula er et lite område av netthinnen som er viktig for skarpsynet. Ved makula-degenerasjon vil skarpsynet hemmes.\n\nTørr makuladegenerasjon gjør at det oppstår små, gule flekker rundt makula. Våt makuladegenerasjon kjennetegnes av små blodkar som vokser på baksiden av øyet.',
+    'Det finnes flere typer, men aldersrelatert makuladegenerasjon er den største årsaken til varig synshemming hos personer over 50 år i Norge. \n\nMakula er et lite område av netthinnen som er viktig for skarpsynet. Ved makula-degenerasjon vil skarpsynet hemmes.\n\nTørr makuladegenerasjon gjør at det oppstår små, gule flekker rundt makula. Våt makuladegenerasjon kjennetegnes av små blodkar som vokser på baksiden av øyet.',
   HEADER_OPPGAVER: 'Oppgaver',
   TEXT_OPPGAVER_1_STAND_1: 'Finn frem mobilen',
   TEXT_OPPGAVER_2_STAND_1: 'Ta på deg simuleringsbrillene',
-  TEXT_OPPGAVER_3_STAND_1: 'Gå inn på yr.no, og finn værvarsel for stedet der du bor',
+  TEXT_OPPGAVER_3_STAND_1:
+    'Gå inn på yr.no, og finn værvarsel for stedet der du bor',
   HEADER_STAND_2: 'Stand 2: Grønn stær',
   TEXT_STAND_2:
-      'Grønn stær er en tilstand som ofte skyldes høyt trykk inne i øyet. På sikt kan tilstanden medføre innskrenkninger i synsfeltet og ubehandlet kan man ende opp med «tunnelsyn». \n\nGlaukom er en av de vanligste årsakene til blindhet i Norge, særlig i høy alder. Vi regner med at 2-3 prosent av befolkningen over 40 år har glaukom, mens antatt forekomst blant de over 70 år er 15-20 prosent.',
+    'Grønn stær er en tilstand som ofte skyldes høyt trykk inne i øyet. På sikt kan tilstanden medføre innskrenkninger i synsfeltet og ubehandlet kan man ende opp med «tunnelsyn». \n\nGlaukom er en av de vanligste årsakene til blindhet i Norge, særlig i høy alder. Vi regner med at 2-3 prosent av befolkningen over 40 år har glaukom, mens antatt forekomst blant de over 70 år er 15-20 prosent.',
   TEXT_OPPGAVER_1_STAND_2: 'Finn frem mobilen',
   TEXT_OPPGAVER_2_STAND_2: 'Ta på deg simuleringsbrillene',
   TEXT_OPPGAVER_3_STAND_2: 'Send en melding på slack eller SMS',
   HEADER_STAND_3: 'Stand 3: Grå stær',
   TEXT_STAND_3:
-      'Grå stær er en sykdom der linsen i øyet mister sin klarhet. Grå stær gir først og fremst nedsatt syn, «tåkesyn», både på avstand og på nært hold.\n\nSelv om alder er den viktigste årsaken til grå stær blir noen barn faktisk født med denne sykdommen.\n\nHos yngre mennesker ses sykdommen oftest på grunn av skader eller andre sykdommer i øyet. Grå stær er den ledende årsaken til blindhet i verden.',
+    'Grå stær er en sykdom der linsen i øyet mister sin klarhet. Grå stær gir først og fremst nedsatt syn, «tåkesyn», både på avstand og på nært hold.\n\nSelv om alder er den viktigste årsaken til grå stær blir noen barn faktisk født med denne sykdommen.\n\nHos yngre mennesker ses sykdommen oftest på grunn av skader eller andre sykdommer i øyet. Grå stær er den ledende årsaken til blindhet i verden.',
   TEXT_OPPGAVER_1_STAND_3: 'Finn frem mobilen',
   TEXT_OPPGAVER_2_STAND_3: 'Ta på deg simuleringsbrillene',
   TEXT_OPPGAVER_3_STAND_3: 'Gå inn på kameraappen og ta en selfie',
   HEADER_STAND_4: 'Stand 4: Tremor (skjelving)',
   TEXT_STAND_4:
-      'Tremor er det samme som skjelving. Dette er en ufrivillig, rytmisk, ofte vedvarende aktivering av muskulaturen.\n\nOfte finner ikke legen årsaken til tremor og den kalles da isolert tremor. Den vanligste formen for isolert tremor er essentiell tremor. Dette er en tremor som rammer hendene spesielt.\n\nI tillegg kan mange også få litt tremor i nakken/hodet og denne kalles da hodetremor.',
+    'Tremor er det samme som skjelving. Dette er en ufrivillig, rytmisk, ofte vedvarende aktivering av muskulaturen.\n\nOfte finner ikke legen årsaken til tremor og den kalles da isolert tremor. Den vanligste formen for isolert tremor er essentiell tremor. Dette er en tremor som rammer hendene spesielt.\n\nI tillegg kan mange også få litt tremor i nakken/hodet og denne kalles da hodetremor.',
   TEXT_OPPGAVER_1_STAND_4: 'Finn frem mobilen',
   TEXT_OPPGAVER_2_STAND_4: 'Ta på deg simuleringshanskene',
   TEXT_OPPGAVER_3_STAND_4: '...',
   HEADER_STAND_5: 'Stand 5: Diabetes retinopati',
   TEXT_STAND_5:
-      'Diabetes retinopati betyr sykdom i øyets netthinne (retina) som følge av diabetes. Dette er den viktigste enkeltårsaken til at personer i yrkesaktiv alder i vestlige land mister synet, helt eller delvis.\n\nPersoner med diabetisk retinopati kan oppleve svekket syn, blanke områder eller blendende lysflekker når de oppholder seg i sterkt lys. Etter som sykdommen utvikler seg, vil uklart syn, blanke områder og svarte flekker opptre. De svarte flekkene vil kunne flyte omkring i synsfeltet.',
+    'Diabetes retinopati betyr sykdom i øyets netthinne (retina) som følge av diabetes. Dette er den viktigste enkeltårsaken til at personer i yrkesaktiv alder i vestlige land mister synet, helt eller delvis.\n\nPersoner med diabetisk retinopati kan oppleve svekket syn, blanke områder eller blendende lysflekker når de oppholder seg i sterkt lys. Etter som sykdommen utvikler seg, vil uklart syn, blanke områder og svarte flekker opptre. De svarte flekkene vil kunne flyte omkring i synsfeltet.',
   TEXT_OPPGAVER_1_STAND_5: 'Finn frem mobilen',
   TEXT_OPPGAVER_2_STAND_5: 'Ta på deg simuleringsbrillene',
   TEXT_OPPGAVER_3_STAND_5: '...',
   HEADER_STAND_6: 'Stand 6: Netthinneløsning',
   TEXT_STAND_6:
-      'Symptomer på netthinneløsning er at synet brått forsvinner. Mange beskriver det som en «gardin» som dras for synsfeltet.\n\nSynet kan gradvis bedre seg i flere måneder etter operasjon. Det blir imidlertid aldri like bra som før. Etter operasjon opplever opp mot halvparten varig synstap.\n\nNedsatt kontrastsyn og forvrengning av synsinntrykk er vanlige følgetilstander som kan gi problemer med lesing, bilkjøring og nærarbeid.',
+    'Symptomer på netthinneløsning er at synet brått forsvinner. Mange beskriver det som en «gardin» som dras for synsfeltet.\n\nSynet kan gradvis bedre seg i flere måneder etter operasjon. Det blir imidlertid aldri like bra som før. Etter operasjon opplever opp mot halvparten varig synstap.\n\nNedsatt kontrastsyn og forvrengning av synsinntrykk er vanlige følgetilstander som kan gi problemer med lesing, bilkjøring og nærarbeid.',
   TEXT_OPPGAVER_1_STAND_6: 'Finn frem mobilen',
   TEXT_OPPGAVER_2_STAND_6: 'Ta på deg simuleringsbrillene',
   TEXT_OPPGAVER_3_STAND_6: '...',
   HEADER_STAND_7: 'Stand 7: Retinitis pigmentosa',
   TEXT_STAND_7:
-      'Retinitis pigmentosa (RP) er fellesbetegnelse for en rekke netthinnesykdommer som gir sterk synshemning i form av innsnevret synsfelt eller blindhet. Sykdommen beskrives som å ha kikkertsyn eller tunnelsyn.\n\nRP er blant de hyppigste årsakene til alvorlig nedsatt syn hos barn og yngre voksne.\n\nSom regel starter RP med at mørkesynet blir dårlig (nattblindhet). Senere er det et stadig mer innsnevret synsfelt (bortfall av sidesyn) som gir de største problemene for mennesker med denne sykdommen.',
+    'Retinitis pigmentosa (RP) er fellesbetegnelse for en rekke netthinnesykdommer som gir sterk synshemning i form av innsnevret synsfelt eller blindhet. Sykdommen beskrives som å ha kikkertsyn eller tunnelsyn.\n\nRP er blant de hyppigste årsakene til alvorlig nedsatt syn hos barn og yngre voksne.\n\nSom regel starter RP med at mørkesynet blir dårlig (nattblindhet). Senere er det et stadig mer innsnevret synsfelt (bortfall av sidesyn) som gir de største problemene for mennesker med denne sykdommen.',
   TEXT_OPPGAVER_1_STAND_7: 'Finn frem mobilen',
   TEXT_OPPGAVER_2_STAND_7: 'Ta på deg simuleringsbrillene',
   TEXT_OPPGAVER_3_STAND_7: '...',
   WARNING_MESSAGE: 'Forsiktighetsregler ved bruk av simuleringshanskene',
-};
+}
 
-export default txt;
+export default txt
